Improve installation advice in configure for Linux
Installing libgrpc-dev is not enough, unfortunately.
I think for some of these packages it is enough the library,
and the `-dev` package is not needed, but then we need to
consider that the library packages have different names
on different distro versions, so it is simpler to list
the `-dev` package.